+++ src/components/design-library-list/design-library-list-item.js
@@ -3,80 +3,64 @@
  */
 import ProControl from '../pro-control'
 import Button from '../button'
-import DEFAULT from './default.json'
-import {
-	addBackgroundScheme, addContainerScheme,
-	addPlaceholderForPostsBlock, cleanParse,
-	getAdditionalStylesForPreview,
-	parseDisabledBlocks,
-} from './util'
+import { DesignPreview } from './design-preview'
+import { useShadowRoot } from './use-shadow-root'
+import { usePreviewRenderer } from './use-preview-renderer'
+import { useAutoScroll } from './use-auto-scroll'
 
 /**
  * External dependencies.
  */
-import { createRoot } from '~stackable/util'
 import { usePresetControls } from '~stackable/hooks'
-import {
-	isPro, i18n, wpGlobalStylesInlineCss, devMode,
-} from 'stackable'
+import { isPro, i18n } from 'stackable'
 import classnames from 'classnames'
 import { Tooltip } from '~stackable/components'
-import { fetchDesign } from '~stackable/design-library'
 
 /**
  * WordPress dependencies.
  */
-import {
-	forwardRef, useEffect, useRef, useState,
-} from '@wordpress/element'
-import { select, useSelect } from '@wordpress/data'
-import { Dashicon } from '@wordpress/components'
+import { forwardRef, useState } from '@wordpress/element'
+import { Dashicon, Spinner } from '@wordpress/components'
 import { __ } from '@wordpress/i18n'
-import { serialize } from '@wordpress/blocks'
-import { applyFilters } from '@wordpress/hooks'
-import { cloneDeep } from 'lodash'
-
-const DEFAULT_CONTENT = { ...DEFAULT }
 
 const DesignLibraryListItem = forwardRef( ( props, ref ) => {
 	const {
-		designId,
 		label,
-		onClick,
-		template = '',
-		category = '',
 		plan,
 		selectedNum = false,
 		selectedData = null,
-		containerScheme,
-		backgroundScheme,
-		enableBackground,
-		cardHeight,
-		setCardHeight,
 		previewSize,
-		setPreviewSize,
+		previewProps,
 		selectedTab,
-		isMultiSelectBusy
+		isMultiSelectBusy,
 	} = props
 
-	const [ parsedBlocks, setParsedBlocks ] = useState( null )
-	const [ content, setContent ] = useState( '' )
+	const presetMarks = usePresetControls( 'spacingSizes' )?.getPresetMarks() || null
 
-	const categoriesRef = useRef( [] )
-	const hostRef = useRef( null )
-	const previewRef = useRef( null )
-	const blocksForSubstitutionRef = useRef( false )
-	const hasBackgroundTargetRef = useRef( false )
-	const initialRenderRef = useRef( null )
-	const scrollPositionRef = useRef( 0 )
-	const shadowBodySizeRef = useRef( null )
+	const spacingSize = ! presetMarks || ! Array.isArray( presetMarks ) ? 120 : presetMarks[ presetMarks.length - 2 ].value
 
-	const scrollIntervalRef = useRef( null )
+	const [ isLoading, setIsLoading ] = useState( true )
 
-	const { getEditorDom } = useSelect( 'stackable/editor-dom' )
-	const editorDom = getEditorDom()
+	const { hostRef, shadowRoot } = useShadowRoot()
 
-	const siteTitle = useSelect( select => select( 'core' ).getEntityRecord( 'root', 'site' )?.title || 'InnovateCo', [] )
+	const {
+		blocks, enableBackground,
+		shadowBodySizeRef, blocksForSubstitutionRef,
+		adjustScale, onClickDesign,
+	} = usePreviewRenderer(
+		previewProps, previewSize, plan, spacingSize,
+		selectedTab, selectedNum, selectedData,
+		ref, shadowRoot, setIsLoading
+	)
+
+	const {
+		onScroll, onMouseOut, onMouseOver,
+	} = useAutoScroll( hostRef, shadowBodySizeRef, selectedTab )
+
+	const getDesignPreviewSize = () => {
+		return selectedNum && selectedData ? selectedData.selectedPreviewSize.preview
+			: ( enableBackground ? previewSize.heightBackground : previewSize.heightNoBackground )
+	}
 
 	const mainClasses = classnames( [
 		'ugb-design-library-item',
@@ -86,344 +70,23 @@ const DesignLibraryListItem = forwardRef( ( props, ref ) => {
 		'ugb--is-toggled': selectedNum,
 	} )
 
-	const STYLE_IDS = applyFilters( 'stackable.global-styles.ids', [
-		'ugb-dep-native-global-style-css-nodep-inline-css',
-		'ugb-style-css-css',
-		'ugb-style-css-responsive-css',
-		'ugb-block-style-inheritance-nodep-inline-css',
-		'ugb-style-css-premium-css',
-	] )
-
-	const addHasBackground = selectedTab === 'patterns'
-	const presetMarks = usePresetControls( 'spacingSizes' )
-		?.getPresetMarks() || null
-
-	const spacingSize = ! presetMarks || ! Array.isArray( presetMarks ) ? 120 : presetMarks[ presetMarks.length - 2 ].value
-
-	const isDesignLibraryDevMode = devMode && localStorage.getItem( 'stk__design_library__dev_mode' ) === '1'
-
-	const adjustPatternSpacing = attributes => {
-		// Note: The styles for block margin/padding will only be generated by the BlockCssGenerator,
-		// which updates the generatedCss attribute. Therefore, updating the block margin/padding
-		// attributes here will have no effect on the preview. Instead, we add additional CSS directly
-		// to the generatedCss attribute for preview purposes. This generatedCss will be regenerated
-		// by the BlockCssGenerator once the pattern is actually added to the editor.
-		let additionalCss = ''
-		if ( ! isDesignLibraryDevMode && category !== 'Header' ) {
-			if ( attributes.hasBackground ) {
-				additionalCss += 'margin-bottom: 0px !important;'
-			} else {
-				additionalCss += `margin-top: ${ spacingSize } !important;`
-				additionalCss += `margin-bottom: ${ spacingSize } !important;`
-			}
-
-			const blockLayouts = select( 'stackable/global-spacing-and-borders' ).getBlockLayouts()
-			if ( attributes.hasBackground && typeof blockLayouts === 'object' && ! blockLayouts[ 'block-background-padding' ] ) {
-				additionalCss += `padding-top: ${ spacingSize } !important;`
-				additionalCss += `padding-right: ${ spacingSize } !important;`
-				additionalCss += `padding-bottom: ${ spacingSize } !important;`
-				additionalCss += `padding-left: ${ spacingSize } !important;`
-			}
-		}
-
-		if ( additionalCss && attributes.uniqueId ) {
-			attributes.generatedCss += `.stk-${ attributes.uniqueId } {${ additionalCss }}`
-		}
-	}
-
-	const adjustScale = () => {
-		if ( ref.current && hostRef.current?.shadowRoot && ! selectedNum ) {
-			const newPreviewSize = { ...previewSize }
-			const newCardHeight = { ...cardHeight }
-			const cardRect = ref.current.getBoundingClientRect()
-
-			const shadowBody = hostRef.current.shadowRoot.querySelector( 'body' )
-			if ( shadowBody ) {
-				const cardWidth = cardRect.width // Get width of the card
-				const scaleFactor = cardWidth > 0 ? cardWidth / shadowBody.offsetWidth : 1 	// Divide by 1200, which is the width of preview in the shadow DOM
-				newPreviewSize.scale = scaleFactor
-
-				const _height = parseFloat( shadowBody.offsetHeight ) * scaleFactor	// Also adjust the height
-
-				if ( Object.keys( newPreviewSize ).length === 1 ) {
-					newPreviewSize.heightBackground = _height
-					newPreviewSize.heightNoBackground = _height
-				} else {
-					const heightKey = enableBackground ? 'heightBackground' : 'heightNoBackground'
-					newPreviewSize[ heightKey ] = _height
-				}
-
-				setPreviewSize( newPreviewSize )
-			}
-
-			if ( ! Object.keys( newCardHeight ).length ) {
-				newCardHeight.background = cardRect.height
-				newCardHeight.noBackground = cardRect.height
-			} else {
-				const CardHeightKey = enableBackground ? 'background' : 'noBackground'
-				newCardHeight[ CardHeightKey ] = cardRect.height
-			}
-
-			setTimeout( () => setCardHeight( newCardHeight ), 500 )
-
-			if ( shadowBodySizeRef.current === null ) {
-				shadowBodySizeRef.current = {
-					clientHeight: shadowBody.clientHeight,
-					scrollHeight: shadowBody.scrollHeight,
-					maxScrollTop: shadowBody.scrollHeight - shadowBody.clientHeight
-				}
-			}
-		}
-	}
-
-	const renderPreview = ( blockContent = content ) => {
-		let blocks = cloneDeep( selectedData?.designData || blockContent )
-
-		// No need to add the color scheme attribute if the selected scheme is the default
-		if ( containerScheme !== '' && ! selectedNum ) {
-			blocks = addContainerScheme( blocks, containerScheme )
-		}
-
-		// Only add a background scheme if it is enabled
-		if ( enableBackground && ! selectedNum ) {
-			blocks = addBackgroundScheme( blocks, enableBackground, backgroundScheme, addHasBackground )
-		}
-
-		if ( selectedTab === 'patterns' ) {
-			adjustPatternSpacing( blocks[ 0 ].attributes )
-		}
-
-		setParsedBlocks( blocks )
-
-		let preview = serialize( blocks )
-
-		// The block `wp/site-title` is a dynamic block, so we need to manually replace it for the preview
-		if ( categoriesRef.current.includes( 'Header' ) ) {
-			preview = preview.replace( /<!--\s*wp:site-title(?:\s+[^\/]*?)?\/-->/g, siteTitle )
-		} else if ( categoriesRef.current.includes( 'Tabs' ) ) {
-		// Add a class for the first tab to be the active tab in the preview
-			preview = preview.replace( '"stk-block-tabs__tab"', '"stk-block-tabs__tab stk-block-tabs__tab--active"' )
-		} else if ( categoriesRef.current.includes( 'Post Loop' ) ) {
-			const defaultValues = DEFAULT_CONTENT[ category ]
-			preview = addPlaceholderForPostsBlock( preview, defaultValues[ 'posts_placeholder' ], defaultValues )
-		}
-
-		const cleanedBlock = preview.replace( /<!--[\s\S]*?-->/g, '' ) // removes comment
-
-		previewRef.current.render( <DesignPreview
-			blocks={ cleanedBlock }
-			adjustScale={ adjustScale }
-			enableBackground={ enableBackground }
-			onScroll={ onScroll }
-		/> )
-	}
-
-	const replacePlaceholders = ( designContent, designCategory ) => {
-		const defaultValues = DEFAULT_CONTENT[ designCategory ]
-
-		if ( defaultValues && ! isDesignLibraryDevMode ) {
-			Object.keys( defaultValues ).forEach( key => {
-				designContent = designContent.replaceAll( key, defaultValues[ key ] )
-			} )
-		}
-
-		return designContent
-	}
-
-	// Replace the placeholders with the default content
-	useEffect( () => {
-		let _blocks = []
-		const initialize = async () => {
-			let _content = template
-			if ( selectedTab === 'patterns' ) {
-				_content = replacePlaceholders( _content, category )
-
-				categoriesRef.current.push( category )
-
-				if ( _content.includes( 'stk-design-library__bg-target="true"' ) ) {
-					hasBackgroundTargetRef.current = true
-				}
-
-				_blocks = cleanParse( _content )
-			} else {
-				for ( let i = 0; i < _content.length; i++ ) {
-					const section = _content[ i ]
-					const design = await fetchDesign( section.id )
-					const designContent = replacePlaceholders( design.template, design.category )
-
-					categoriesRef.current.push( design.category )
-
-					let _block = cleanParse( designContent )[ 0 ]
-
-					if ( section.bg ) {
-						_block = addBackgroundScheme( [ _block ], true, '' )[ 0 ]
-					}
-
-					adjustPatternSpacing( _block.attributes )
-					_blocks.push( _block )
-				}
-			}
-		}
-
-		initialize().then( () => {
-			// We need to parse the content because this is what we use to insert the blocks in the Editor
-			// const _block = cleanParse( _content )[ 0 ]
-			const { blocks, blocksForSubstitution } = parseDisabledBlocks( _blocks )
-			blocksForSubstitutionRef.current = blocksForSubstitution
-
-			const shadowRoot = hostRef.current.shadowRoot || hostRef.current.attachShadow( { mode: 'open' } )
-
-			// Initialize the shadow DOM for the first time
-			if ( ! previewRef.current ) {
-				// Get all styles needed and make a copy for the shadow DOM
-				const styleNodes = STYLE_IDS.map( id => {
-					let style = null
-					let node = document?.head?.querySelector( `#${ id }` )
-					if ( ! node && editorDom ) {
-						const editorBody = editorDom?.closest( 'body' )
-						const editorHead = editorBody?.ownerDocument?.head
-						node = editorHead.querySelector( `#${ id }` )
-					}
-
-					if ( node ) {
-						style = node.cloneNode( true )
-					}
-
-					return style
-				} ).filter( node => node !== null )
-
-				// Add global and theme styles
-				const globalStylesNode = document.createElement( 'style' )
-				globalStylesNode.setAttribute( 'id', 'global-styles-inline-css' )
-				globalStylesNode.innerHTML = wpGlobalStylesInlineCss
-				styleNodes.push( globalStylesNode )
-
-				const hostStyles = document.createElement( 'style' )
-				hostStyles.setAttribute( 'id', 'stk-design-library-styles' )
-
-				// Additional styles for blocks to render properly in the preview
-				hostStyles.innerHTML += getAdditionalStylesForPreview()
-
-				styleNodes.push( hostStyles )
-
-				styleNodes.forEach( node => {
-					if ( node.textContent ) {
-						// we use :host in the shadow DOM to target the root
-						node.textContent = node.textContent.replace( /:root/g, ':host' )
-					}
-					shadowRoot.appendChild( node )
-				} )
-
-				previewRef.current = createRoot( shadowRoot )
-			}
-
-			renderPreview( blocks )
-			setContent( blocks )
-		} )
-	}, [ template ] )
-
-	useEffect( () => {
-		if ( ! initialRenderRef.current ) {
-			initialRenderRef.current = true
+	const onClickHost = e => {
+		e.stopPropagation()
+		if ( selectedTab === 'pages' ) {
 			return
 		}
-
-		if ( ! content ||
-			! previewRef.current ||
-			selectedNum
-		) {
-			return
-		}
-
-		renderPreview()
-	}, [ content, containerScheme, backgroundScheme, enableBackground ] )
-
-	useEffect( () => {
-		if ( selectedNum === 0 && content && previewRef.current ) {
-			renderPreview()
-			adjustScale()
-		}
-	}, [ selectedNum ] )
-
-	const getDesignPreviewSize = () => {
-		return selectedNum && selectedData ? selectedData.selectedPreviewSize.preview
-			: ( enableBackground ? previewSize.heightBackground : previewSize.heightNoBackground )
-	}
-
-	const onClickDesign = () => {
-		if ( ! isPro && plan !== 'free' ) {
-			return
-		}
-		const cardRect = ref.current.getBoundingClientRect()
-
-		const selectedPreviewSize = {
-			preview: enableBackground ? previewSize.heightBackground : previewSize.heightNoBackground,
-			card: cardRect.height,
-			scale: previewSize.scale,
-		}
-
-		onClick( designId, category, parsedBlocks, blocksForSubstitutionRef.current, selectedPreviewSize )
-	}
-
-	const onMouseOver = () => {
-		const shadowDomBody = hostRef?.current?.shadowRoot.querySelector( 'body' )
-		if ( shadowDomBody && shadowBodySizeRef.current ) {
-			setTimeout( () => {
-				if ( scrollIntervalRef.current ) {
-					clearInterval( scrollIntervalRef.current )
-				}
-
-				if ( scrollPositionRef.current === -1 ) {
-					scrollPositionRef.current = shadowDomBody.scrollTop
-				}
-
-				scrollIntervalRef.current = setInterval( () => {
-					console.log( 'scrolling' )
-					const scrollDifference = shadowBodySizeRef.current.maxScrollTop - scrollPositionRef.current
-					const shouldScroll = shadowBodySizeRef.current.maxScrollTop - scrollPositionRef.current > 0
-
-					if ( shadowDomBody && shouldScroll ) {
-						const scrollBy = scrollDifference >= 20 ? 20 : scrollDifference
-						shadowDomBody.scrollTop = scrollPositionRef.current + scrollBy
-						scrollPositionRef.current += scrollBy
-					} else {
-						clearInterval( scrollIntervalRef.current )
-					}
-				}, 20 )
-			}, 500 )
-		}
-	}
-
-	const onMouseOut = () => {
-		if ( scrollPositionRef.current <= 0 ) {
-			return
-		}
-
-		const shadowDomBody = hostRef?.current?.shadowRoot.querySelector( 'body' )
-		if ( shadowDomBody ) {
-			clearInterval( scrollIntervalRef.current )
-			shadowDomBody.scrollTo({
-				top: 0,
-				behavior: 'smooth'
-			})
-			scrollPositionRef.current = -1
-		}
-	}
-
-	const onScroll = () => {
-		if ( scrollIntervalRef.current ) {
-			clearInterval( scrollIntervalRef.current )
-		}
+		onClickDesign()
 	}
 
 	return (
+		// eslint-disable-next-line jsx-a11y/mouse-events-have-key-events
 		<button
 			className={ mainClasses }
 			ref={ ref }
 			data-stk-design-id={ props.designId }
-			onClick={ () => selectedTab === 'patterns' ? onClickDesign() : {} }
-			onMouseOut={ () => selectedTab === 'patterns' ? {} : onMouseOut() }
-			onMouseOver={ () => selectedTab === 'patterns' ? {} : onMouseOver() }
+			onClick={ onClickHost }
+			onMouseOut={ onMouseOut }
+			onMouseOver={ onMouseOver }
 		>
 			{ ! isPro && plan !== 'free' && <span className="stk-pulsating-circle" role="presentation" /> }
 			<div style={ { position: 'relative' } } className={ `${ getDesignPreviewSize() > 100 ? 'stk--design-preview-large' : 'stk--design-preview-small' }` }>
@@ -434,6 +97,7 @@ const DesignLibraryListItem = forwardRef( ( props, ref ) => {
 						showHideNote={ false }
 					/>
 				) }
+				{ isLoading && <Spinner style={ { display: 'block', margin: '50px auto' } } /> }
 				<div
 					className="stk-block-design__host-container"
 					style={ {
@@ -443,8 +107,15 @@ const DesignLibraryListItem = forwardRef( ( props, ref ) => {
 							: ( enableBackground ? previewSize.heightBackground : previewSize.heightNoBackground ),
 					} }
 				>
-
-					<div className="stk-block-design__host" ref={ hostRef } />
+					<div className="stk-block-design__host" ref={ hostRef }>
+						{ shadowRoot && <DesignPreview
+							blocks={ blocks }
+							shadowRoot={ shadowRoot }
+							selectedTab={ selectedTab }
+							adjustScale={ adjustScale }
+							onScroll={ onScroll }
+						/> }
+					</div>
 				</div>
 			</div>
 
@@ -467,17 +138,17 @@ const DesignLibraryListItem = forwardRef( ( props, ref ) => {
 						</Tooltip>
 					}
 					{ selectedTab === 'patterns' ? <span className="stk-block-design__selected-num">{ selectedNum === 0 ? '' : selectedNum }</span>
-					: <div>
-						<Button
-							label={ __( 'Insert', i18n ) }
-							className="ugb-modal-design-library__add-multi"
-							disabled={ isMultiSelectBusy }
-							onClick={ () => onClickDesign() }
-						>
-							{ __( 'Insert', i18n ) }
-							{ isMultiSelectBusy && <Spinner /> }
-						</Button>
-					</div>
+						: <div>
+							<Button
+								label={ __( 'Insert', i18n ) }
+								className="ugb-modal-design-library__add-multi"
+								disabled={ isMultiSelectBusy }
+								onClick={ () => onClickDesign() }
+							>
+								{ __( 'Insert', i18n ) }
+								{ isMultiSelectBusy && <Spinner /> }
+							</Button>
+						</div>
 					}
 				</div>
 			</footer>
@@ -495,27 +166,3 @@ DesignLibraryListItem.defaultProps = {
 }
 
 export default DesignLibraryListItem
-
-const DesignPreview = ( {
-	blocks, adjustScale, enableBackground, onScroll
-} ) => {
-	useEffect( () => {
-		// Adjust scale if the background was toggled
-		adjustScale()
-		setTimeout( adjustScale, 50 )
-	}, [ blocks, enableBackground ] )
-
-	const shadowBodyClasses = classnames( applyFilters( 'stackable.global-styles.classnames', [ 'entry-content' ] ) )
-
-	return (
-		<body
-			className={ shadowBodyClasses }
-			onWheel={ () => onScroll() }
-		>
-			<div
-				dangerouslySetInnerHTML={ { __html: blocks } }
-				style={ { pointerEvents: 'none' } }	// prevent blocks from being clicked
-			/>
-		</body>
-	)
-}
