+++ src/components/design-library-list/util.js
@@ -1,9 +1,12 @@
 /* eslint-disable no-console */
 import DEFAULT from './default.json'
+
 import {
 	settings, isPro, cdnUrl,
 } from 'stackable'
+
 import { parse, serialize } from '@wordpress/blocks'
+import { select } from '@wordpress/data'
 
 const DEFAULT_CONTENT = { ...DEFAULT }
 const PARSER = new DOMParser()
@@ -147,19 +150,45 @@ export const addBackgroundScheme = (
 	} )
 }
 
-export const replacePlaceholders = ( block, category ) => {
-	const key = block.attributes.text
-	const defaultValues = DEFAULT_CONTENT[ category ]
-	if ( key && defaultValues ) {
-		const newValue = defaultValues[ key ]
-		block.attributes.text = newValue
+export const adjustPatternSpacing = ( attributes, category, spacingSize, isDesignLibraryDevMode ) => {
+	// Note: The styles for block margin/padding will only be generated by the BlockCssGenerator,
+	// which updates the generatedCss attribute. Therefore, updating the block margin/padding
+	// attributes here will have no effect on the preview. Instead, we add additional CSS directly
+	// to the generatedCss attribute for preview purposes. This generatedCss will be regenerated
+	// by the BlockCssGenerator once the pattern is actually added to the editor.
+	let additionalCss = ''
+	if ( ! isDesignLibraryDevMode && category !== 'Header' ) {
+		if ( attributes.hasBackground ) {
+			additionalCss += 'margin-bottom: 0px !important;'
+		} else {
+			additionalCss += `margin-top: ${ spacingSize } !important;`
+			additionalCss += `margin-bottom: ${ spacingSize } !important;`
+		}
+
+		const blockLayouts = select( 'stackable/global-spacing-and-borders' ).getBlockLayouts()
+		if ( attributes.hasBackground && typeof blockLayouts === 'object' && ! blockLayouts[ 'block-background-padding' ] ) {
+			additionalCss += `padding-top: ${ spacingSize } !important;`
+			additionalCss += `padding-right: ${ spacingSize } !important;`
+			additionalCss += `padding-bottom: ${ spacingSize } !important;`
+			additionalCss += `padding-left: ${ spacingSize } !important;`
+		}
 	}
 
-	block.innerBlocks.forEach( innerBlock => {
-		replacePlaceholders( innerBlock, category )
-	} )
+	if ( additionalCss && attributes.uniqueId ) {
+		attributes.generatedCss += `.stk-${ attributes.uniqueId } {${ additionalCss }}`
+	}
+}
 
-	return block
+export const replacePlaceholders = ( designContent, designCategory, isDesignLibraryDevMode ) => {
+	const defaultValues = DEFAULT_CONTENT[ designCategory ]
+
+	if ( defaultValues && ! isDesignLibraryDevMode ) {
+		Object.keys( defaultValues ).forEach( key => {
+			designContent = designContent.replaceAll( key, defaultValues[ key ] )
+		} )
+	}
+
+	return designContent
 }
 
 const retrieveAttributes = ( blockName, hasNoInnerBlocks, content ) => {
@@ -183,7 +212,7 @@ export const parseDisabledBlocks = parsedBlocks => {
 	const blocksForSubstitution = new Set()
 
 	if ( Array.isArray( disabledBlocks ) && ! disabledBlocks.length ) {
-		return { blocks: parsedBlocks, blocksForSubstitution }
+		return [ parsedBlocks, blocksForSubstitution ]
 	}
 
 	const addOriginalAttributes = blocks => {
@@ -203,7 +232,7 @@ export const parseDisabledBlocks = parsedBlocks => {
 	}
 
 	const blocks = addOriginalAttributes( parsedBlocks )
-	return { blocks, blocksForSubstitution }
+	return [ blocks, blocksForSubstitution ]
 }
 
 const IMAGE_STORAGE = cdnUrl.replace( /\/$/, '' ) + '/library-v4/images/'
@@ -268,10 +297,10 @@ export const getAdditionalStylesForPreview = () => {
 	// Creates a Block Formatting Context (BFC), ensuring it contains the margin of its children
 	styles += 'body { display: flow-root; }'
 
-	styles += 'body { --stk-block-wide-width: 1000px; max-height: 1500px;}'
+	styles += 'body { --stk-block-wide-width: 1000px; }'
 
 	// Use 'auto' for scroll behavior. Let the scroll interval handle the speed.
-	styles += 'body { scroll-behavior: auto; overflow-y: auto; }'
+	styles += 'body.preview-pages { scroll-behavior: auto; overflow-y: auto; max-height: 1500px; }'
 
 	styles += '::-webkit-scrollbar { width: 0; }'
